modifying ndimage filters for new interface

+++ cudipy/segment/tests/test_mask.py
@@ -15,13 +15,14 @@ from cudipy.segment.mask import (
 )
 from dipy.data import get_fnames
 from dipy.io.image import load_nifti_data
-from cupyx.scipy.ndimage.filters import median_filter
+from cupyx.scipy.ndimage import median_filter
 from cupyx.scipy.ndimage import binary_dilation, generate_binary_structure
 
 
-@cp.testing.with_requires('cupyimg')
+@cp.testing.with_requires("cupyimg")
 def test_mask():
     from cudipy.segment.mask import otsu
+
     vol = cp.zeros((30, 30, 30))
     vol[15, 15, 15] = 1
     struct = generate_binary_structure(3, 1)
@@ -93,11 +94,11 @@ def test_bounding_box():
         assert_equal(maxs, [0, 0])
 
 
-@cp.testing.with_requires('cupyimg')
+@cp.testing.with_requires("cupyimg")
 def test_median_otsu():
-    fname = get_fnames('S0_10')
+    fname = get_fnames("S0_10")
     data = load_nifti_data(fname)
-    data = cp.asarray(np.squeeze(data.astype('f8')))
+    data = cp.asarray(np.squeeze(data.astype("f8")))
     dummy_mask = data > data.mean()
     data_masked, mask = median_otsu(
         data,
